Attempted fix for systems where sizeof(int) < sizeof(void*)
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@21770 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSCallBacks.m
@@ -67,7 +67,7 @@ _NS_int_release(void *table, void* i)
 NSString *
 _NS_int_describe(void *table, void* i)
 {
-  return [NSString stringWithFormat: @"%d", i];
+  return [NSString stringWithFormat: @"%d", (int)i];
 }
 
 /** For owned `void *' **/
